% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transCCA.R
\name{transCCA}
\alias{transCCA}
\title{Output transelliptical CCA directions and correlations}
\usage{
transCCA(x, y, ndir, eigenmin = 0.001)
}
\arguments{
\item{x}{The first data matrix to be included in the calculation of the transformed Kendall scatter matrix}

\item{y}{The second data matrix to be included in the calculation of the transformed Kendall scatter matrix}

\item{ndir}{The number of canonical directions to estimate}

\item{eigenmin}{The minimum eigenvalue when transforming matrix to be positive definite}
}
\value{
The estimate of transelliptical canonical correlations and directions
}
\description{
This takes two data sets and returns the transelliptical canonical directions and correlations
}
