% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transCorVar.R
\name{transCorVar}
\alias{transCorVar}
\title{Transelliptical Canonical correlation asymptotic variance}
\usage{
transCorVar(x, y, i, eigenmin = 0.001)
}
\arguments{
\item{x}{The first data matrix to be included in the calculation of the transformed Kendall scatter matrix}

\item{y}{The second data matrix to be included in the calculation of the transformed Kendall scatter matrix}

\item{i}{The index for the canonical correlation variance to estimate}

\item{eigenmin}{The minimum eigenvalue when transforming matrix to be positive definite}
}
\value{
The estimate of the asymptotic variance for the ith transelliptical canonical correlation and direction
}
\description{
Estimates the asymptotic variance for the ith transelliptical canonical correlation and direction estimate
}
